// src/app/modules/shared/models/product.interface.ts
// Generated by https://quicktype.io

export interface Product {
    id:          number;
    title:       string;
    price:       number;
    description: string;
    category:    string;
    image:       string;
    rating:      Rating;
}

export interface Rating {
    rate:  number;
    count: number;
}